Overhaul/Addition of Animation System
No longer will we use the idea of an AnimatedSpriteComponent to dictate what resource a SpriteComponent should display (as for what worked before "the refactor"). Now we use an extendsion of the Sprite, AnimatedSprite, that knows about what states and resourceIDs to load when updated by the AnimationSystem. TODO: Implement a State machine to work with this and incorporate ResourceID Struct.